backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-filter==23.3
django-ratelimit==4.1.0
python-dotenv==1.0.0
requests==2.31.0
huggingface_hub==0.22.2
mysqlclient==2.2.0
mysql-connector-python==8.2.0
sqlparse>=0.4.4
cryptography==41.0.3
openpyxl==3.1.2
numpy==1.23.5
pandas==2.0.3
httpx==0.27.0
cffi==1.16.0
docxtpl==0.16.4
imapclient==3.0.1
email-validator==2.0.0
pytz==2023.3
python-dateutil==2.8.2
